drivers/led: Extend userled_set_t from 8bits to 32bits

+++ boards/arm/cxd56xx/spresense/src/cxd56_userleds.c
@@ -88,7 +88,7 @@ void board_userled(int led, bool ledon)
  * Name: board_userled_all
  ****************************************************************************/
 
-void board_userled_all(uint8_t ledset)
+void board_userled_all(uint32_t ledset)
 {
   cxd56_gpio_write(GPIO_LED1, (ledset & BOARD_LED1_BIT) != 0);
   cxd56_gpio_write(GPIO_LED2, (ledset & BOARD_LED2_BIT) != 0);
