+++ boards/arm/nrf52/nrf52840-dongle/src/nrf52_userleds.c
@@ -65,7 +65,7 @@
  *
  *  uint32_t board_userled_initialize(void);
  *  void board_userled(int led, bool ledon);
- *  void board_userled_all(uint8_t ledset);
+ *  void board_userled_all(uint32_t ledset);
  */
 
 #define LED_ON 0
@@ -150,7 +150,7 @@ void board_userled(int led, bool ledon)
  * Name: board_userled_all
  ****************************************************************************/
 
-void board_userled_all(uint8_t ledset)
+void board_userled_all(uint32_t ledset)
 {
   int i;
