+++ boards/arm/stm32h7/stm32h747i-disco/src/stm32_userleds.c
@@ -132,7 +132,7 @@ void board_userled(int led, bool ledon)
  *
  ****************************************************************************/
 
-void board_userled_all(uint8_t ledset)
+void board_userled_all(uint32_t ledset)
 {
   /* Active Low */
